Move common utils
