+++ client/packages/requisitions/src/common/index.tsx
@@ -2,3 +2,4 @@ export * from './ListIndicators';
 export * from './PageLayout';
 export * from './ModalContentPanels';
 export * from './ModalContentLayout';
+export * from './utils';
